Add files via upload
Data Article
Data mining for the study of the Epidemic (SARS-CoV-2) COVID-19: Algorithm for the identification of patients speaking the native language in the Totonacapan area – Mexico

Authors
Carlos Medel-Ramírez 1, Hilario Medel-López 2

Affiliations
1.	Universidad Veracruzana / Instituto de Investigaciones y Estudios Superiores Económicos y Sociales
2.	Universidad Veracruzana / Instituto de Antropología

Corresponding author
Carlos Medel-Ramírez (cmedel@uv.mx)

Abstract
The importance of the working document is that it allows analyzing the information and the status of the cases associated with (SARS-CoV-2) COVID-19 as data open to the municipal government and especially in the Totonacapan Zone in Mexico, from the registry patient diary, according to age, sex, comorbidities and condition of (SARS-CoV-2) COVID-19, according to the following characteristics: a) Positive, b) Negative, c) Suspect. Likewise, it presents information on the identification of an outpatient and / or hospitalized patient, attending to their medical development, identifying: a) Recovered, b) Deaths and c) Assets. Data analysis is carried out by applying a data mining algorithm, which provides the information, fast and timely, necessary for the estimation of the healthcare scenarios of (SARS-CoV-2) COVID-19.

Keywords
(SARS-CoV-2) COVID-19, Algorithm (SARS-CoV-2) COVID-19, Totonacapan Zone, Mexico, Identification of patients, Native language